move succesfully deprecation helpers to util to organize the code
