+++ segregation/aspatial/aspatial_indexes.py
@@ -11,6 +11,8 @@ import warnings
 from scipy.stats import norm
 from scipy.optimize import minimize
 
+from segregation.util.util import _dep_message, DeprecationHelper
+
 # Including old and new api in __all__ so users can use both
 
 __all__ = ['Dissim', 
@@ -1819,34 +1821,7 @@ class DensityCorrectedDissim:
 
 
 
-# Deprecation Calls (_dep_message and DeprecationHelper could be moved to some utility class) #
-# However, this was atempted, but I was crashing due to circular calls #
-        
-def _dep_message(original, replacement, when="2020-01-31", version="2.1.0"):
-    msg = "Deprecated (%s): %s" % (version, original)
-    msg += " is being renamed to %s." % replacement
-    msg += " %s will be removed on %s." % (original, when)
-    return msg
-
-class DeprecationHelper(object):
-    def __init__(self, new_target, message="Deprecated"):
-        self.new_target = new_target
-        self.message = message
-
-    def _warn(self):
-        from warnings import warn
-
-        warn(self.message)
-
-    def __call__(self, *args, **kwargs):
-        self._warn()
-        return self.new_target(*args, **kwargs)
-
-    def __getattr__(self, attr):
-        self._warn()
-        return getattr(self.new_target, attr)
-        
-
+# Deprecation Calls
         
 msg = _dep_message("Gini_Seg", "GiniSeg")
 Gini_Seg = DeprecationHelper(GiniSeg, message=msg)
