+++ segregation/aspatial/multigroup_aspatial_indexes.py
@@ -5,8 +5,9 @@ Multigroup Aspatial based Segregation Metrics
 __author__ = "Renan X. Cortes <renanc@ucr.edu>, Sergio J. Rey <sergio.rey@ucr.edu> and Elijah Knaap <elijah.knaap@ucr.edu>"
 
 import numpy as np
-import pandas as pd
 from sklearn.metrics.pairwise import manhattan_distances
+
+from segregation.util.util import _dep_message, DeprecationHelper
 
 # Including old and new api in __all__ so users can use both
 
@@ -1143,35 +1144,8 @@ class MultiDivergence:
 
 
 
-# Deprecation Calls (_dep_message and DeprecationHelper could be moved to some utility class) #
-# However, this was atempted, but I was crashing due to circular calls #
-        
-def _dep_message(original, replacement, when="2020-01-31", version="2.1.0"):
-    msg = "Deprecated (%s): %s" % (version, original)
-    msg += " is being renamed to %s." % replacement
-    msg += " %s will be removed on %s." % (original, when)
-    return msg
+# Deprecation Calls
 
-class DeprecationHelper(object):
-    def __init__(self, new_target, message="Deprecated"):
-        self.new_target = new_target
-        self.message = message
-
-    def _warn(self):
-        from warnings import warn
-
-        warn(self.message)
-
-    def __call__(self, *args, **kwargs):
-        self._warn()
-        return self.new_target(*args, **kwargs)
-
-    def __getattr__(self, attr):
-        self._warn()
-        return getattr(self.new_target, attr)
-        
-
-        
 msg = _dep_message("Multi_Dissim", "MultiDissim")
 Multi_Dissim = DeprecationHelper(MultiDissim, message=msg)
