+++ segregation/decomposition/decompose_segregation.py
@@ -6,7 +6,7 @@ __author__ = "Renan X. Cortes <renanc@ucr.edu>, Elijah Knaap <elijah.knaap@ucr.e
 
 
 import warnings
-from segregation.util.util import _generate_counterfactual
+from segregation.util.util import _generate_counterfactual, _dep_message, DeprecationHelper
 
 # Including old and new api in __all__ so users can use both
 
@@ -299,34 +299,7 @@ class DecomposeSegregation:
 
 
 
-# Deprecation Calls (_dep_message and DeprecationHelper could be moved to some utility class) #
-# However, this was atempted, but I was crashing due to circular calls #
-        
-def _dep_message(original, replacement, when="2020-01-31", version="2.1.0"):
-    msg = "Deprecated (%s): %s" % (version, original)
-    msg += " is being renamed to %s." % replacement
-    msg += " %s will be removed on %s." % (original, when)
-    return msg
+# Deprecation Calls
 
-class DeprecationHelper(object):
-    def __init__(self, new_target, message="Deprecated"):
-        self.new_target = new_target
-        self.message = message
-
-    def _warn(self):
-        from warnings import warn
-
-        warn(self.message)
-
-    def __call__(self, *args, **kwargs):
-        self._warn()
-        return self.new_target(*args, **kwargs)
-
-    def __getattr__(self, attr):
-        self._warn()
-        return getattr(self.new_target, attr)
-        
-
-        
 msg = _dep_message("Decompose_Segregation", "DecomposeSegregation")
 Decompose_Segregation = DeprecationHelper(DecomposeSegregation, message=msg)
