+++ segregation/local/local_indexes.py
@@ -12,6 +12,8 @@ import libpysal as lps
 
 from segregation.spatial import RelativeCentralization
 
+from segregation.util.util import _dep_message, DeprecationHelper
+
 # Including old and new api in __all__ so users can use both
 
 __all__ = [
@@ -718,35 +720,8 @@ class LocalRelativeCentralization:
 
 
 
-# Deprecation Calls (_dep_message and DeprecationHelper could be moved to some utility class) #
-# However, this was atempted, but I was crashing due to circular calls #
-        
-def _dep_message(original, replacement, when="2020-01-31", version="2.1.0"):
-    msg = "Deprecated (%s): %s" % (version, original)
-    msg += " is being renamed to %s." % replacement
-    msg += " %s will be removed on %s." % (original, when)
-    return msg
+# Deprecation Calls
 
-class DeprecationHelper(object):
-    def __init__(self, new_target, message="Deprecated"):
-        self.new_target = new_target
-        self.message = message
-
-    def _warn(self):
-        from warnings import warn
-
-        warn(self.message)
-
-    def __call__(self, *args, **kwargs):
-        self._warn()
-        return self.new_target(*args, **kwargs)
-
-    def __getattr__(self, attr):
-        self._warn()
-        return getattr(self.new_target, attr)
-        
-
-        
 msg = _dep_message("Multi_Location_Quotient", "MultiLocationQuotient")
 Multi_Location_Quotient = DeprecationHelper(MultiLocationQuotient, message=msg)
